test: test duplicated user try to sign up
